pyproject.toml
[tool.poetry]
name = "aws-terraform-sandbox"
version = "0.1.0"
description = ""
authors = ["Maki <sunwood.ai.labs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
markdown = "^3.6"
jinja2 = "^3.1.4"
rich = "^13.7.1"
art = "^6.2"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
